Add roles to users

+++ src/main/java/com/crojas/demo/domain/UserDto.java
@@ -5,10 +5,10 @@ import lombok.Data;
 @Data
 public class UserDto {
 
-    private int userId;
+    private Long userId;
     private String firstName, lastName, username;
 
-    public UserDto(int userId, String firstName, String lastName, String username) {
+    public UserDto(Long userId, String firstName, String lastName, String username) {
         this.userId = userId;
         this.firstName = firstName;
         this.lastName = lastName;
